# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Proof Of Learn API
  version: 1.0.0
  description: The REST API for Proof Of Learn platform.
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /projects:
    get:
      summary: List all projects
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: difficulty
          schema:
            type: string
      responses:
        '200':
          description: Successful response
  /projects/{id}:
    get:
      summary: Get project details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
...